This is (some) of the Gracenote stuff.
I took out the problem columns (mood, tempo, and genre being some of them). If there's anything useful to work with here, you can go ahead and start using it. I'm in the process of fixing those problem columns, and I'll add them to the data set and upload new versions of the data as I add things back in.